+++ src/test/testData/contract/AdminMixinIndex/mixin.js
@@ -1,0 +1,7 @@
+Mixin.register('contract-mixin', {
+    methods: {
+        contractHelper() {
+            return 'contract';
+        },
+    },
+});