+++ src/test/testData/contract/ScriptHookIndex/ContractHook.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types=1);
+
+namespace Contract\Script;
+
+use Shopware\Core\Framework\Script\Execution\Hook;
+
+class ContractHook extends Hook
+{
+    public const HOOK_NAME = 'contract.hook';
+
+    public static function getServiceIds(): array
+    {
+        return [];
+    }
+}