+++ src/test/testData/contract/ShopwareBundleIndex/ContractBundle/src/ContractBundle.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Contract\Bundle;
+
+use Shopware\Core\Framework\Plugin;
+
+class ContractBundle extends Plugin
+{
+}