+++ src/test/testData/isolation/projectA/administration/shared-component.js
@@ -1,0 +1,8 @@
+Component.register('sw-shared-isolation', {
+    props: {
+        alphaProp: {
+            type: String,
+            required: false,
+        },
+    },
+});